// src/enums/brews/brewAction.ts
export enum BREW_ACTION {
  POST = 'POST',
  REPEAT = 'REPEAT',
  DETAIL = 'DETAIL',
  EDIT = 'EDIT',
  DELETE = 'DELETE',
  PHOTO_GALLERY = 'PHOTO_GALLERY',
  CUPPING = 'CUPPING',
  SHOW_MAP_COORDINATES = 'SHOW_MAP_COORDINATES',
  FAST_REPEAT = 'FAST_REPEAT',
  TOGGLE_FAVOURITE = 'TOGGLE_FAVOURITE',
  SHARE = 'SHARE',
  VISUALIZER = 'VISUALIZER',
  RATING = 'RATING',
  TOGGLE_BEST_BREW = 'TOGGLE_BEST_BREW',
  SHOW_VISUALIZER = 'SHOW_VISUALIZER',
  SHOW_GRAPH = 'SHOW_GRAPH',
  DOWNLOAD_XLSX = 'DOWNLOAD_XLSX',
}